Extract base class logic into a mixin
