+++ src/api-viewer-base.ts
@@ -1,26 +1,9 @@
 import { LitElement, html, property, TemplateResult } from 'lit-element';
 import { until } from 'lit-html/directives/until.js';
-import { ElementInfo, ElementSetInfo } from './lib/types.js';
+import { ElementPromise } from './lib/types.js';
 import { setTemplates } from './lib/utils.js';
+import { ApiViewerMixin } from './api-viewer-mixin.js';
 import './api-viewer-content.js';
-
-type ElementPromise = Promise<ElementInfo[]>;
-
-async function fetchJson(src: string): ElementPromise {
-  let result: ElementInfo[] = [];
-  try {
-    const file = await fetch(src);
-    const json = (await file.json()) as ElementSetInfo;
-    if (Array.isArray(json.tags) && json.tags.length) {
-      result = json.tags;
-    } else {
-      console.error(`No element definitions found at ${src}`);
-    }
-  } catch (e) {
-    console.error(e);
-  }
-  return result;
-}
 
 async function renderDocs(
   jsonFetched: ElementPromise,
@@ -52,40 +35,20 @@ async function renderDocs(
 
 let id = 0;
 
-export class ApiViewerBase extends LitElement {
-  @property({ type: String }) src?: string;
-
+export class ApiViewerBase extends ApiViewerMixin(LitElement) {
   @property({ type: String }) section = 'docs';
 
-  @property({ type: String }) selected?: string;
-
   @property({ type: String, attribute: 'exclude-knobs' }) excludeKnobs?: string;
-
-  @property({ attribute: false })
-  elements?: ElementInfo[];
-
-  private jsonFetched: ElementPromise = Promise.resolve([]);
-
-  private lastSrc?: string;
 
   protected _id?: number;
 
   constructor() {
     super();
 
-    this._id = ++id;
+    this._id = id++;
   }
 
   protected render() {
-    const { src } = this;
-
-    if (Array.isArray(this.elements)) {
-      this.jsonFetched = Promise.resolve(this.elements);
-    } else if (src && this.lastSrc !== src) {
-      this.lastSrc = src;
-      this.jsonFetched = fetchJson(src);
-    }
-
     return html`
       ${until(
         renderDocs(
